Store host attributes alongside offers to reduce number of lookups.
Bugs closed: AURORA-913

Reviewed at https://reviews.apache.org/r/27902/

+++ src/main/java/org/apache/aurora/scheduler/async/RandomJitterReturnDelay.java
@@ -43,7 +43,7 @@ class RandomJitterReturnDelay implements OfferReturnDelay {
   }
 
   @Override
-  public Amount<Integer, Time> get() {
-    return Amount.of(minHoldTimeMs + random.nextInt(maxJitterWindowMs), Time.MILLISECONDS);
+  public Amount<Long, Time> get() {
+    return Amount.of((long) minHoldTimeMs + random.nextInt(maxJitterWindowMs), Time.MILLISECONDS);
   }
 }
